Added Conversions header/cpp
